organised the git structure
